# target/classes/openapi.yaml
openapi: 3.0.0
info:
  title: Title
  description: Title
  version: 1.0.0
servers:
  - url: 'https'
paths:
  /topsecret:
    post:
      summary: gets the location and message of the ship
      operationId: getDataShip
      requestBody:
        description: satellites
        required: true
        content:
          application/json:
            schema:
              type: object
              title: shipDataRequest
              properties:
                satellites:
                  type: array
                  items:
                    $ref: '#/components/schemas/SatelliteRequest'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: object
                title: shipDataResponse
                properties:
                  position:
                    $ref: '#/components/schemas/Coordinate'
                  message:
                    type: string
        404:
          description: ERROR The message or the possition couldn't be determinated
          content:
            application/json:
              schema:
                type: object
                title: shipDataResponse
                properties:
                  position:
                    $ref: '#/components/schemas/Coordinate'
                  message:
                    type: string
  /topsecret_split:
    get:
      summary: gets the location and message of the ship
      operationId: getSplitDataShip
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: object
                title: shipDataResponse
                properties:
                  position:
                    $ref: '#/components/schemas/Coordinate'
                  message:
                    type: string
        404:
          description: ERROR The message or the possition couldn't be determinated
          content:
            application/json:
              schema:
                type: object
                title: shipDataResponse
                properties:
                  position:
                    $ref: '#/components/schemas/Coordinate'
                  message:
                    type: string
  /topsecret_split/{satellite_name}:
    post:
      summary: update message and location from satellite
      operationId: updateSplitSatelliteData
      parameters:
        - name: satellite_name
          in: path
          description: satellite to update
          required: true
          schema:
            type: string
      requestBody:
        description: satellite
        required: true
        content:
          application/json:
            schema:
              type: object
              title: splitShipDataRequest
              properties:
                distance:
                  type: number
                  format: float
                message:
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: object
                title: shipDataResponse
                properties:
                  position:
                    $ref: '#/components/schemas/Coordinate'
                  message:
                    type: string
        404:
          description: ERROR The message or the possition couldn't be determinated
          content:
            application/json:
              schema:
                type: object
                title: shipDataResponse
                properties:
                  position:
                    $ref: '#/components/schemas/Coordinate'
                  message:
                    type: string
  
components:
  schemas:
    SatelliteRequest:
      type: object
      properties:
        name:
          type: string
        distance:
          type: number
          format: float
        message:
          type: array
          items:
            type: string
    Coordinate:
      type: object
      properties:
        x:
          type: number
          format: float
        y:
          type: number
          format: float